requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
gunicorn
python-dotenv
pymysql